Update to upstream version '0.0.23'
with Debian dir d7d7e1463b8a020d0cefbf4471ee8cbec66076de

+++ swh/loader/tar/loader.py
@@ -1,9 +1,9 @@
-# Copyright (C) 2015-2016  The Software Heritage developers
+# Copyright (C) 2015-2017  The Software Heritage developers
 # See the AUTHORS file at the top-level directory of this distribution
 # License: GNU General Public License version 3, or any later version
 # See top-level LICENSE file for more information
 
-import datetime
+
 import os
 import tempfile
 import shutil
@@ -14,34 +14,23 @@ from swh.loader.tar import tarball, utils
 
 
 class TarLoader(loader.DirLoader):
-    """A tarball loader.
+    """A tarball loader:
 
-    """
-    CONFIG_BASE_FILENAME = 'loader/tar'
-
-    ADDITIONAL_CONFIG = {
-        'extraction_dir': ('string', '/tmp')
-    }
-
-    def __init__(self):
-        super().__init__(logging_class='swh.loader.tar.TarLoader')
-
-    def load(self, tarpath, origin, visit, revision, release, occurrences):
-        """
-        Load a tarball in backend.
-
-        This will:
+        - creates an origin if it does not exist
+        - creates a fetch_history entry
+        - creates an origin_visit
         - uncompress locally the tarballs in a temporary location
         - process the content of the tarballs to persist on swh storage
         - clean up the temporary location
         - write an entry in fetch_history to mark the loading tarball end
+        (success or failure)
 
         Args:
             - tarpath: path to the tarball to uncompress
             - origin: Dictionary origin
               - url: url origin we fetched
               - type: type of the origin
-            - visit: Numbered visit
+            - visit_date (str): To override the visit date
             - revision: Dictionary of information needed, keys are:
               - author_name: revision's author name
               - author_email: revision's author email
@@ -53,20 +42,32 @@ class TarLoader(loader.DirLoader):
               - committer_offset: date offset e.g. -0220, +0100
               - type: type of revision dir, tar
               - message: synthetic message for the revision
-            - release: Dictionary of information needed, keys are:
-              - name: release name
-              - date: release timestamp (e.g. 1444054085)
-              - offset: release date offset e.g. -0220, +0100
-              - author_name: release author's name
-              - author_email: release author's email
-              - comment: release's comment message
             - occurrences: List of occurrence dictionary.
               Information needed, keys are:
               - branch: occurrence's branch name
               - authority_id: authority id (e.g. 1 for swh)
               - validity: validity date (e.g. 2015-01-01 00:00:00+00)
 
+    """
+    CONFIG_BASE_FILENAME = 'loader/tar'
+
+    ADDITIONAL_CONFIG = {
+        'extraction_dir': ('string', '/tmp')
+    }
+
+    def __init__(self):
+        super().__init__(logging_class='swh.loader.tar.TarLoader')
+
+    def prepare(self, *args, **kwargs):
+        """1. Uncompress the tarball in a temporary directory.
+           2. Compute some metadata to update the revision.
+
         """
+        tarpath, origin, visit_date, revision, occs = args
+
+        if 'type' not in origin:  # let the type flow if present
+            origin['type'] = 'tar'
+
         # Prepare the extraction path
         extraction_dir = self.config['extraction_dir']
         os.makedirs(extraction_dir, 0o755, exist_ok=True)
@@ -77,54 +78,23 @@ class TarLoader(loader.DirLoader):
         artifact = utils.convert_to_hex(hashutil.hashfile(tarpath))
         artifact['name'] = os.path.basename(tarpath)
 
-        try:
-            self.log.info('Uncompress %s to %s' % (tarpath, dir_path))
-            nature = tarball.uncompress(tarpath, dir_path)
-            artifact['archive_type'] = nature
-            artifact['length'] = os.path.getsize(tarpath)
+        self.log.info('Uncompress %s to %s' % (tarpath, dir_path))
+        nature = tarball.uncompress(tarpath, dir_path)
+        artifact['archive_type'] = nature
+        artifact['length'] = os.path.getsize(tarpath)
 
-            revision['metadata'] = {
-                'original_artifact': [artifact],
-            }
+        revision['metadata'] = {
+            'original_artifact': [artifact],
+        }
 
-            return super().load(
-                dir_path, origin, visit, revision, release, occurrences)
-        finally:
+        self.dir_path = dir_path
+
+        super().prepare(dir_path, origin, visit_date, revision, None, occs)
+
+    def cleanup(self):
+        """Clean up temporary directory where we uncompress the tarball.
+
+        """
+        dir_path = self.dir_path
+        if dir_path and os.path.exists(dir_path):
             shutil.rmtree(dir_path)
-
-    def prepare_and_load(self,
-                         tarpath, origin, revision, release, occurrences):
-        """
-        Prepare origin, fetch_origin, origin_visit
-        Then load a tarball 'tarpath'.
-        Then close origin_visit, fetch_history
-
-        First:
-        - creates an origin if it does not exist
-        - creates a fetch_history entry
-        - creates an origin_visit
-        - Then loads the tarball
-
-        """
-        if 'type' not in origin:  # let the type flow if present
-            origin['type'] = 'tar'
-
-        self.origin_id = self.storage.origin_add_one(origin)
-        origin['id'] = self.origin_id
-
-        date_visit = datetime.datetime.now(tz=datetime.timezone.utc)
-        origin_visit = self.storage.origin_visit_add(origin['id'], date_visit)
-        visit = origin_visit['visit']
-
-        fetch_history_id = self.open_fetch_history()
-
-        try:
-            self.load(tarpath, origin, visit, revision, release, occurrences)
-            self.close_fetch_history_success(fetch_history_id)
-            self.storage.origin_visit_update(
-                self.origin_id, origin_visit['visit'], status='full')
-        except:
-            self.close_fetch_history_failure(fetch_history_id)
-            self.storage.origin_visit_update(
-                self.origin_id, origin_visit['visit'], status='partial')
-            raise
